apps/hetzner-m365-rag/api/requirements.txt
# FastAPI and server
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Elasticsearch
elasticsearch==8.15.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# MinIO (S3-compatible storage)
minio==7.2.3

# Pydantic
pydantic==2.6.1
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# OpenAI
openai==1.12.0

# Microsoft Graph / M365
msal==1.26.0
httpx==0.26.0

# RAG-Anything (installed separately in Dockerfile)
# raganything[all]

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3

# Logging
structlog==24.1.0

